proj0 fix bug 2

+++ proj0/NBody.java
@@ -1,7 +1,7 @@
 public class NBody {
     public static double readRadius(String path){
         In in = new In(path);        
-        int  num_planet = in.readDouble();
+        int num_planet = in.readInt();
         double radius = in.readDouble();
         return radius;
     }
